requirements.txt
# --- Core Framework ---
fastapi==0.115.0
uvicorn==0.30.1
python-multipart==0.0.9

# --- Database ---
motor==3.3.1        # Async MongoDB driver
pymongo==4.6.3

# --- Security & Auth ---
PyJWT==2.9.0
passlib[bcrypt]==1.7.4

# --- ML / NLP ---
langchain==0.3.2
langchain-community==0.3.1
transformers==4.45.1
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1

# --- PDF & DOCX Processing ---
PyPDF2==3.0.1 
python-docx==1.1.2

# --- Utilities ---
pydantic==2.9.2
pydantic-settings==2.6.1
numpy==1.26.4

# --- Optional (if using .env) ---
python-dotenv==1.0.1
